requirements.txt
streamlit
opencv-python
face-alignment
pillow
numpy
torch
torchvision
scikit-learn
